doctype_creator/requirements.txt
# DocType Creator Requirements

# Core dependencies
PyYAML>=6.0
jsonschema>=4.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
